🚚 Rename app, package and theme
