+++ android/app/src/androidTest/java/live/ditto/reactinmobile/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package live.ditto.host
+package live.ditto.reactinmobile
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
