+++ android/app/src/main/java/live/ditto/reactinmobile/ComposeWrappedWebView.kt
@@ -1,4 +1,4 @@
-package live.ditto.host
+package live.ditto.reactinmobile
 
 // The built in Android WebView
 import android.webkit.WebResourceRequest
@@ -18,7 +18,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.platform.LocalInspectionMode
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.viewinterop.AndroidView
-import live.ditto.host.ui.theme.HostTheme
+import live.ditto.reactinmobile.ui.theme.ReactInMobileTheme
 
 @Composable
 fun ComposeWrappedWebView() {
@@ -76,7 +76,7 @@ fun ComposeWrappedWebView() {
 @Preview(showBackground = true, apiLevel = 33)
 @Composable
 fun ComposeWrappedWebViewPreview() {
-    HostTheme {
+    ReactInMobileTheme {
         ComposeWrappedWebView()
     }
 }
