+++ android/app/src/main/java/live/ditto/reactinmobile/MainActivity.kt
@@ -1,4 +1,4 @@
-package live.ditto.host
+package live.ditto.reactinmobile
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -7,13 +7,13 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import live.ditto.host.ui.theme.HostTheme
+import live.ditto.reactinmobile.ui.theme.ReactInMobileTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            HostTheme {
+            ReactInMobileTheme {
                 // A surface container using the 'background' color from the theme
                 Surface(
                     modifier = Modifier.fillMaxSize(),
@@ -25,4 +25,3 @@ class MainActivity : ComponentActivity() {
         }
     }
 }
-
