+++ android/app/src/main/java/live/ditto/reactinmobile/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package live.ditto.host.ui.theme
+package live.ditto.reactinmobile.ui.theme
 
 import android.app.Activity
 import android.os.Build
@@ -38,7 +38,7 @@ private val LightColorScheme = lightColorScheme(
 )
 
 @Composable
-fun HostTheme(
+fun ReactInMobileTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
     // Dynamic color is available on Android 12+
     dynamicColor: Boolean = true,
