+++ android/app/src/main/java/live/ditto/reactinmobile/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package live.ditto.host.ui.theme
+package live.ditto.reactinmobile.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
